Updating auth and fixing bug
