+++ src/controllers/auth.controller.ts
@@ -90,19 +90,6 @@ export const login = async (req: Request, res: Response, next: NextFunction) => 
         })(req, res, next); 
 }
 
-export const authenticateJWT = (req: Request, res: Response, next: NextFunction) => {
-    passport.authenticate('jwt', { session: false }, (err, user, info) => {
-      if (err) {
-        return next(err);
-      }
-      if (!user) {
-        return next(new UnauthorizedError('Not authorized to access this resource🤔'));
-      }
-      
-      next();
-    })(req, res, next);
-};
-
 export const google = async (req: Request, res: Response, next: NextFunction) => {
     passport.authenticate('google', {
         accessType: 'offline',
